+++ chat-widget/src/common/telemetry/TelemetryHelper.ts
@@ -274,4 +274,17 @@ export class TelemetryHelper {
         };
         BroadcastService.postMessage(telemetryEvent);
     }
+
+    public static logWebChatEvent = (logLevel: LogLevel, payload: TelemetryEventWrapper) => {
+        const telemetryEvent: ITelemetryEvent = {
+            eventName: TelemetryEvent.WebChatEvent,
+            logLevel: logLevel,
+            payload: {
+                ...payload,
+                type: TelemetryEvent.WebChatEvent,
+                scenarioType: ScenarioType.WEBCHAT
+            } as WebChatTelemetryData
+        };
+        BroadcastService.postMessage(telemetryEvent);
+    }
 }