+++ chat-widget/src/components/livechatwidget/common/defaultProps/dummyDefaultProps.ts
@@ -15,6 +15,7 @@ import { createAvatarMiddleware } from "../../../webchatcontainerstateful/webcha
 import { createMarkdown } from "../createMarkdown";
 import { groupActivitiesMiddleware } from "../../../webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/groupActivitiesMiddleware";
 import { typingIndicatorMiddleware } from "../../../webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/typingIndicatorMiddleware";
+import { createWebChatTelemetry } from "../../../webchatcontainerstateful/webchatcontroller/webchattelemetry/WebChatLogger";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export const dummyDefaultProps: ILiveChatWidgetProps = {
@@ -1445,7 +1446,7 @@ export const dummyDefaultProps: ILiveChatWidgetProps = {
             internalErrorBoxClass: undefined,
             internalRenderErrorBox: undefined,
             locale: "en-US",
-            onTelemetry: undefined,
+            onTelemetry: createWebChatTelemetry(),
             overrideLocalizedStrings: undefined,
             renderMarkdown: createMarkdown(false, false),
             scrollToEndButtonMiddleware: undefined,