+++ CHANGE_LOG.md
@@ -15,3 +15,4 @@ All notable changes to this project will be documented in this file.## [Unreleas
 - Added broadcast event support for startChat and endChat
 - Added custom componentOverride for preChat
 - Added Telemetry events for PostChat Loading Pane
+- Added web chat telemetry logger